move models.py into new directory
