+++ backend/app/models/__init__.py
@@ -1,0 +1,21 @@
+from .inquiries import Inquiry, InquiryPublic, InquiryCreate
+
+from .deprecated_models import (
+    Item,
+    ItemCreate,
+    ItemUpdate,
+    ItemPublic,
+    ItemsPublic,
+    User,
+    UserCreate,
+    UserUpdate,
+    UserPublic,
+    UsersPublic,
+    UserRegister,
+    UserUpdateMe,
+    Message,
+    NewPassword,
+    UpdatePassword,
+    Token,
+    TokenPayload,
+)
